move: offer - routers
