+++ app/entities/offer/admins_offer.py
