+++ app/entities/offer/schemas_offer.py
