+++ app/includer/admins.py
@@ -1,5 +1,5 @@
 from entities.category.admins_category import CategoryAdmin
-from entities.offer.admin_offer import OfferLinkAdmin
+from entities.offer.admins_offer import OfferLinkAdmin
 from entities.service_name.admins_service_name import ServiceNameAdmin
 from entities.users.admins_user import EmployeeAdmin, ClientAdmin
